resources: 앱 아이콘 리소스 업데이트 및 좌표 표시 형식 수정
- 캔버스 오프셋 좌표를 정수형으로 표시하도록 변경
- 신규 앱 아이콘 및 플레이 스토어용 이미지 리소스 추가 및 업데이트

+++ app/src/main/java/com/boostcamp/and03/ui/screen/canvasmemo/component/CanvasStatusCard.kt
@@ -70,8 +70,8 @@ fun CanvasStatusCard(
                 Text(
                     text = stringResource(
                         R.string.canvas_offset_format,
-                        canvasViewOffset.x,
-                        canvasViewOffset.y
+                        canvasViewOffset.x.toInt(),
+                        canvasViewOffset.y.toInt()
                     ),
                     style = MaterialTheme.typography.bodySmall,
                     color = And03Theme.colors.onSurface
